requirements.txt
apache-airflow==2.3.0
requests==2.31.0
pandas==2.0.3
pyarrow==12.0.0
pytest==7.4.0